fix vc7.x warnings
git-svn-id: http://svn.boost.org/svn/boost/trunk@26435 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/date_time/posix_time/conversion.hpp
@@ -24,7 +24,7 @@ namespace posix_time {
   ptime from_time_t(std::time_t t) 
   {
     ptime start(gregorian::date(1970,1,1));
-    return start + seconds(t);
+    return start + seconds(static_cast<long>(t));
   }
 
   //! Convert a time to a tm structure truncating any fractional seconds
